# src/DataGovIL.Client/Docs/Swagger.JSON
{
  "openapi": "3.0.1",
  "info": {
    "title": "DataGovIL.Api",
    "description": "Sample .NET Core Web API on top of the data.gov.il CKAN datastore (vehicle registrations + WLTP makes/models).",
    "version": "v1"
  },
  "paths": {
    "/api/Manufacturers": {
      "get": {
        "tags": [
          "Manufacturers"
        ],
        "summary": "Paginated list of car manufacturers and models (from the WLTP makes/models dataset).",
        "parameters": [
          {
            "name": "manufacturer",
            "in": "query",
            "description": "Optional exact manufacturer name filter (e.g. \"טויוטה יפן\").",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model",
            "in": "query",
            "description": "Optional commercial model-name search (e.g. \"COROLLA\" also matches \"COROLLA CROSS\").",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "1-based page number.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Rows per page (max 500).",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ManufacturerModelRecordPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManufacturerModelRecordPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManufacturerModelRecordPagedResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/Manufacturers/{manufacturerCode}/{modelCode}": {
      "get": {
        "tags": [
          "Manufacturers"
        ],
        "summary": "Look up a single make/model by manufacturer code (`tozeret_cd`) and model code (`degem_cd`).",
        "parameters": [
          {
            "name": "manufacturerCode",
            "in": "path",
            "description": "e.g. \"5\".",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "modelCode",
            "in": "path",
            "description": "e.g. \"179\".",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ManufacturerModelRecord"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManufacturerModelRecord"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManufacturerModelRecord"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Manufacturers/export": {
      "post": {
        "tags": [
          "Manufacturers"
        ],
        "summary": "Queues a background export of the full WLTP make/model catalog to data.csv\r\n(one row per unique tozeret_cd + degem_cd). Returns immediately.",
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ManufacturerExportStatus"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManufacturerExportStatus"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManufacturerExportStatus"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ManufacturerExportStatus"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManufacturerExportStatus"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManufacturerExportStatus"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Vehicles/{registrationNumber}/history": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Technical + ownership history for a registration number\r\n(`mispar_rechev`) from the two history datastore resources.",
        "parameters": [
          {
            "name": "registrationNumber",
            "in": "path",
            "description": "e.g. \"12345678\".",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleHistoryRecord"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleHistoryRecord"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleHistoryRecord"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Vehicles/{registrationNumber}": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Look up a single vehicle by its exact registration number (מספר רישוי).\r\nWhen the active registries miss, falls back to permanently-cancelled (\"ביטול סופי\")\r\nresources and then to inactive (\"לא פעיל\") resources, then personal-import vehicles.\r\nA found vehicle is also checked\r\nagainst the safety-systems discount list and the open-recall list, and gets a\r\ndepreciation calculation (kilometers, owner count, ownership type) from its history and price list.",
        "parameters": [
          {
            "name": "registrationNumber",
            "in": "path",
            "description": "e.g. \"12345678\".",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleRecord"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleRecord"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleRecord"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Vehicles": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Paged/free-text search across the vehicle registration dataset.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleRecordPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleRecordPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleRecordPagedResult"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ManufacturerExportStatus": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "pagesFetched": {
            "type": "integer",
            "format": "int32"
          },
          "recordsSeen": {
            "type": "integer",
            "format": "int32"
          },
          "uniqueRowCount": {
            "type": "integer",
            "format": "int32"
          },
          "estimatedTotalRawRows": {
            "type": "integer",
            "description": "CKAN's include_total figure from the first page. The WLTP resource currently returns\r\nan estimate (`total_was_estimated`), so this is progress context only — not a\r\nstop condition.",
            "format": "int32",
            "nullable": true
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "outputFile": {
            "type": "string",
            "description": "Destination of the last successful sync (e.g. `postgresql:public.manufacturer_models`).\r\nKept as `OutputFile` for API compatibility with the former CSV export.",
            "nullable": true
          },
          "inserted": {
            "type": "integer",
            "description": "Totals across all synced tables; see DataGovIL.Api.Models.ManufacturerExportStatus.Tables for the per-table breakdown.",
            "format": "int32",
            "nullable": true
          },
          "updated": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "deleted": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "unchanged": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "tables": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/ManufacturerSyncDelta"
            },
            "description": "Delta per table, keyed by table name (e.g. `manufacturers`, `manufacturer_models`).",
            "nullable": true
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Snapshot of the in-process manufacturer CSV export job (single-instance deployment)."
      },
...
      "ManufacturerModelRecordPagedResult": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "description": "Total number of matching rows, when known (datastore include_total).",
            "format": "int32",
            "nullable": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ManufacturerModelRecord"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ManufacturerSyncDelta": {
        "type": "object",
        "properties": {
          "inserted": {
            "type": "integer",
            "format": "int32"
          },
          "updated": {
            "type": "integer",
            "format": "int32"
          },
          "deleted": {
            "type": "integer",
            "format": "int32"
          },
          "unchanged": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Counts from a manufacturer catalog delta sync."
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "VehicleHistoryRecord": {
        "type": "object",
        "properties": {
          "registrationNumber": {
            "type": "string",
            "nullable": true
          },
          "technical": {
            "$ref": "#/components/schemas/VehicleTechnicalHistoryRecord"
          },
          "ownershipHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VehicleOwnershipHistoryRecord"
            },
            "description": "Ownership timeline rows (typically multiple), newest first when sortable.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Combined vehicle history from the technical and ownership datastore resources,\r\nkeyed by registration number (`mispar_rechev`)."
      },
...
      "VehicleRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "מזהה",
            "nullable": true
          },
          "registrationNumber": {
            "type": "string",
            "description": "מספר רכב",
            "nullable": true
          },
          "manufacturerCode": {
            "type": "string",
            "description": "קוד תוצר",
            "nullable": true
          },
          "modelType": {
            "type": "string",
            "description": "סוג דגם",
            "nullable": true
          },
          "manufacturerName": {
            "type": "string",
            "description": "תוצר",
            "nullable": true
          },
          "modelCode": {
            "type": "string",
            "description": "קוד דגם",
            "nullable": true
          },
          "modelName": {
            "type": "string",
            "description": "דגם",
            "nullable": true
          },
          "vehicleTypeCode": {
            "type": "string",
            "description": "קוד סוג רכב",
            "nullable": true
          },
          "vehicleTypeName": {
            "type": "string",
            "description": "סוג רכב",
            "nullable": true
          },
          "trimLevel": {
            "type": "string",
            "description": "רמת גימור",
            "nullable": true
          },
          "safetyEquipmentLevel": {
            "type": "string",
            "description": "רמת אבזור בטיחותי",
            "nullable": true
          },
          "pollutionGroup": {
            "type": "string",
            "description": "קבוצת זיהום",
            "nullable": true
          },
          "manufactureYear": {
            "type": "string",
            "description": "שנת יצור",
            "nullable": true
          },
          "engineManufacturer": {
            "type": "string",
            "description": "תוצר מנוע",
            "nullable": true
          },
          "engineModel": {
            "type": "string",
            "description": "דגם מנוע",
            "nullable": true
          },
          "engineNumber": {
            "type": "string",
            "description": "מספר מנוע",
            "nullable": true
          },
          "totalWeight": {
            "type": "string",
            "description": "משקל כולל",
            "nullable": true
          },
          "lastTestDate": {
            "type": "string",
            "description": "תאריך מבחן אחרון",
            "nullable": true
          },
          "testValidUntil": {
            "type": "string",
            "description": "תוקף",
            "nullable": true
          },
          "cancellationDate": {
            "type": "string",
            "description": "תאריך ביטול",
            "nullable": true
          },
          "ownershipType": {
            "type": "string",
            "description": "בעלות",
            "nullable": true
          },
          "chassisNumber": {
            "type": "string",
            "description": "מספר שילדה",
            "nullable": true
          },
          "colorCode": {
            "type": "string",
            "description": "קוד צבע",
            "nullable": true
          },
          "color": {
            "type": "string",
            "description": "צבע רכב",
            "nullable": true
          },
          "frontTire": {
            "type": "string",
            "description": "צמיג קדמי",
            "nullable": true
          },
          "rearTire": {
            "type": "string",
            "description": "צמיג אחורי",
            "nullable": true
          },
          "fuelType": {
            "type": "string",
            "description": "סוג דלק",
            "nullable": true
          },
          "registrationOrder": {
            "type": "string",
            "description": "הוראת רישום",
            "nullable": true
          },
          "roadEntryDate": {
            "type": "string",
            "description": "מועד עליה לכביש",
            "nullable": true
          },
          "estimatedWarrantyEndDate": {
            "type": "string",
            "format": "date",
            "description": "סיום אחריות יצרן משוערת: roadEntryDate + 36 months. Null without a road entry date.",
            "nullable": true,
            "readOnly": true
          },
          "estimatedWarrantyActive": {
            "type": "boolean",
            "description": "אחריות יצרן משוערת בתוקף. Null without a road entry date.",
            "nullable": true,
            "readOnly": true
          },
          "updatedDate": {
            "type": "string",
            "description": "תאריך עדכון",
            "nullable": true
          },
          "importType": {
            "type": "string",
            "description": "סוג יבוא",
            "nullable": true
          },
          "commercialName": {
            "type": "string",
            "description": "כינוי מסחרי",
            "nullable": true
          },
          "isPermanentlyCancelled": {
            "type": "boolean",
            "description": "ביטול סופי"
          },
          "isInactive": {
            "type": "boolean",
            "description": "לא פעיל"
          },
          "isSafetyDiscountEligible": {
            "type": "boolean",
            "description": "זכאי להנחה לאחר התקנת מערכות בטיחות"
          },
          "manufacturerModel": {
            "$ref": "#/components/schemas/ManufacturerModelRecord"
          },
          "recalls": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VehicleRecallRecord"
            },
            "description": "ריקולים",
            "nullable": true
          },
          "history": {
            "$ref": "#/components/schemas/VehicleHistoryRecord"
          },
          "depreciation": {
            "$ref": "#/components/schemas/VehicleDepreciationRecord"
          },
          "extensionData": {
            "type": "object",
            "additionalProperties": { },
            "description": "שדות נוספים",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Public API shape for a vehicle registration row (active, inactive, or permanently cancelled).\r\nSerialized with ASP.NET's default camelCase names.\r\nBuilt from DataGovIL.Api.Models.VehicleDatastoreRecord, DataGovIL.Api.Models.VehicleCancelledDatastoreRecord,\r\nDataGovIL.Api.Models.VehicleInactiveWithoutModelCodeDatastoreRecord, DataGovIL.Api.Models.VehicleSafetyDiscountDatastoreRecord,\r\nor DataGovIL.Api.Models.VehiclePersonalImportDatastoreRecord."
      },
      "VehicleRecordPagedResult": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "description": "Total number of matching rows, when known (datastore include_total).",
            "format": "int32",
            "nullable": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VehicleRecord"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}